+++ gems/aws-sdk-core/lib/seahorse/client/net_http/handler.rb
@@ -42,7 +42,13 @@ module Seahorse
         # @param [RequestContext] context
         # @return [Response]
         def call(context)
-          transmit(context.config, context.http_request, context.http_response)
+          span_wrapper(context) do
+            transmit(
+              context.config,
+              context.http_request,
+              context.http_response
+            )
+          end
           Response.new(context: context)
         end
 
@@ -192,6 +198,17 @@ module Seahorse
           end
         end
 
+        def span_wrapper(context, &block)
+          context.tracer.in_span(
+            'Handler.NetHttp',
+            attributes: Aws::Telemetry.http_request_attrs(context)
+          ) do |span|
+            block.call
+            span.add_attributes(
+              Aws::Telemetry.http_response_attrs(context)
+            )
+          end
+        end
       end
     end
   end