+++ gems/aws-sdk-core/spec/spec_helper.rb
@@ -70,6 +70,16 @@ module SpecHelper
       client_class.new(options)
     end
 
+    # clears opentelemetry-sdk configuration state between specs
+    # https://github.com/open-telemetry/opentelemetry-ruby/blob/main/test_helpers/lib/opentelemetry/test_helpers.rb#L18
+    def reset_opentelemetry_sdk
+      OpenTelemetry.instance_variable_set(
+        :@tracer_provider,
+        OpenTelemetry::Internal::ProxyTracerProvider.new
+      )
+      OpenTelemetry.error_handler = nil
+      OpenTelemetry.propagation = nil
+    end
   end
 end